Add --n-workers option (#252)
Add switching the number of parallelism (the default value is the number
of online processors in Ruby 2.2.0 or later, but fixed them to 1 for old
Ruby). Please note that the `Thread` based runner is not yet available
(raises `NameError`).

Note:

We considered the number of parallelism option names below:

1. `--n-threads`
    * Backend is not just about `Thread`
    * Seems does not make sense
2. `--jobs`
    * We don't call this a job
    * Seems does not make sense
3. `--n-workers`
    * We already call this a worker
    * Seems make sense

for future parallelization support. Part of GH-235.

---------

Co-authored-by: Sutou Kouhei <kou@clear-code.com>