Show the top 5 slow tests with test source location
GitHub: GH-253

This is a part of adding support for showing the top 5 slow tests in the
summary output.

This just show the top 5 slow tests with test source location for easy
to jump in the editor. So it doesn't show the following information in
the summary output for now.

- Command lines (easy to re-run)