Show the top 5 slow tests with --location option
GitHub: fix GH-253

This is a **final** part of adding support for showing the top 5 slow
tests in the summary output.

This show the top 5 slow tests with `--location` option for easy to
re-run command lines.

Removing the `=` to make it easier to jump in the editor. Because
including the `=` makes it difficult to jump in the editor:

`--location ...` is better than
`--location=...`
           ^