test: use require_relative instead of require
GitHub: fix GH-311

This is a follow-up to GH-312 and the final patch to fix GH-311.
This patch will reduce dependency on `$LOAD_PATH`.

It replaces `require` with `require_relative` for internal library
files in `test/`.

If you want to run tests against the installed `test-unit`, you can
still use the `test/` directory like this:

This example runs tests against the installed `test-unit` version 3.6.8.

```console
$ ruby -I $(gem env home)/gems/test-unit-3.6.8/lib bin/test-unit test
```